// frontend/src/features/actividades/verificar_correccion_escritura/models/verificar_correccion_escritura.ts
export interface VerificarCorreccionEscritura {
  id_verificar_correccion_escritura: number;
  id_expediente: number;
  observaciones?: string | null;
  is_active: boolean;
  row_status: boolean;
  created_by: number;
  created_date: string;
  modified_by?: number | null;
  modified_date?: string | null;
}